penubra bugfix
For low-resolution profiles like the StarCheck sometimes the 80% or 20% position were not calculated correctly because of selecting the wrong range in the data, leading to a clearly wrong penumbra value. This (rather old) part of the code is completely reviewed and improved. Even on extreme sparse data it will function as expected now.
Minor bug: When switching to the next item in the history list, the reference was lost when there was not set a temporary reference. A reload was the work-around. This is fixed now.
Code review: the multi-threaded Nelder-Mead code was redesigned as singular engine with shared data such that multiple independent engines can run in their own thread. Progress reporting is now sent asynchronously by the main thread, thus avoiding many multi-threading complications. Procedure calls were optimised for maximum efficiency.
Code review: some memory leaks due to not freed structures before closing the program were resolved.
Reducing the size of the history list will also unset the temporary reference if it is no longer available.