Wedge alignment bug and refactoring
Minor bug fix: Even with Wedge detection on, alignment could lead to a wrong center position for very wide wedged fields and resulting in unreliable alignment for the reference. This could be corrected with a reload of the profile. Now the edge found through the derivative is used when for the derivative is concluded that a wedged profile is found.

Slightly faster analysis by avoiding recalculation of cpu-intensive results if already available.

A CPU load indicator and a Field Type have been added to the statusbar.

The functionality of the Processing option Mirrored measurement as Reference is extended to all profiles in the history list.